+++ Api/NotificationCenter/Services/NotificationService.cs
@@ -1,143 +1,106 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using CSharpFunctionalExtensions;
+using HappyTravel.Edo.Api.Models.Agents;
+using HappyTravel.Edo.Api.Models.Mailing;
+using HappyTravel.Edo.Api.Models.Users;
+using HappyTravel.Edo.Api.NotificationCenter.Services;
+using HappyTravel.Edo.Api.Services.Agents;
+using HappyTravel.Edo.Common.Enums;
 using HappyTravel.Edo.Notifications.Enums;
 using HappyTravel.Edo.Notifications.Models;
-using HappyTravel.Edo.Data;
-using Microsoft.EntityFrameworkCore;
-using HappyTravel.Edo.Api.NotificationCenter.Hubs;
-using HappyTravel.Edo.Api.NotificationCenter.Models;
-using Microsoft.AspNetCore.SignalR;
+using System.Collections.Generic;
 using System.Text.Json;
-using HappyTravel.Edo.Api.Infrastructure;
+using System.Threading.Tasks;
 
 namespace HappyTravel.Edo.Api.NotificationCenter.Services
 {
     public class NotificationService : INotificationService
     {
-        public NotificationService(EdoContext context, 
-            IHubContext<AgentNotificationHub, INotificationClient> agentNotificationHub,
-            IHubContext<AdminNotificationHub, INotificationClient> adminNotificationHub,
-            MailSenderWithCompanyInfo mailSender,
-            IDateTimeProvider dateTimeProvider)
+        public NotificationService(IInternalNotificationService internalNotificationService,
+            INotificationOptionsService notificationOptionsService,
+            IAgentContextService agentContextService)
         {
-            _context = context;
-            _agentNotificationHub = agentNotificationHub;
-            _adminNotificationHub = adminNotificationHub;
-            _mailSender = mailSender;
-            _dateTimeProvider = dateTimeProvider;
+            _internalNotificationService = internalNotificationService;
+            _notificationOptionsService = notificationOptionsService;
+            _agentContextService = agentContextService;
         }
-        
 
-        public async Task Add(Notifications.Models.Notification notification)
+
+        public async Task<Result> Send(ApiCaller apiCaller, JsonDocument message, NotificationTypes notificationType, string email = "", string templateId = "")
         {
-            var entry = _context.Notifications.Add(new Data.Notifications.Notification
+            return await Send(apiCaller, message, notificationType, new List<string> { email }, templateId);
+        }
+
+
+        public async Task<Result> Send(ApiCaller apiCaller, JsonDocument message, NotificationTypes notificationType, List<string> emails = null, string templateId = "")
+        {
+            if (apiCaller.Type == ApiCallerTypes.Agent)
             {
-                Receiver = notification.Receiver,
-                UserId = notification.UserId,
-                AgencyId = notification.AgencyId,
-                Message = notification.Message,
-                Type = notification.Type,
-                SendingSettings = JsonDocument.Parse(JsonSerializer.SerializeToUtf8Bytes(notification.SendingSettings, new(JsonSerializerDefaults.Web))),
-                Created = _dateTimeProvider.UtcNow()
-            });
-            await _context.SaveChangesAsync();
+                var agent = await _agentContextService.GetAgent();
 
-            var tasks = new List<Task>();
-
-            foreach (var (protocol, settings) in notification.SendingSettings)
-            {
-                var task = protocol switch
-                {
-                    ProtocolTypes.Email when settings is EmailSettings emailSettings 
-                        => SendEmail(emailSettings),
-                    
-                    ProtocolTypes.WebSocket when settings is WebSocketSettings webSocketSettings
-                        => notification.Receiver switch 
-                        {
-                            ReceiverTypes.AgentApp
-                                => SendMessageToAgent(notification.UserId, notification.AgencyId, entry.Entity.Id, notification.Message),
-                            
-                            ReceiverTypes.AdminPanel
-                                => SendMessageToAdmin(notification.UserId, entry.Entity.Id, notification.Message),
-
-                            _ => throw new ArgumentException($"Unsupported receiver '{notification.Receiver}' for notification")
-                        },
-                    
-                    _ => throw new ArgumentException($"Unsupported protocol '{protocol}' or incorrect settings type")
-                };
-                
-                tasks.Add(task);
+                return await Send(new SlimAgentContext(agent.AgentId, agent.AgencyId), message, notificationType, emails, templateId);
             }
-
-            await Task.WhenAll(tasks);
-        }
-
-
-        public async Task MarkAsRead(int notificationId)
-        {
-            var notification = await _context.Notifications
-                .SingleOrDefaultAsync(n => n.Id == notificationId && !n.IsRead);
-
-            if (notification is not null)
-            {
-                notification.IsRead = true;
-                _context.Notifications.Update(notification);
-                await _context.SaveChangesAsync();
-            }
-        }
-
-
-        public async Task<List<SlimNotification>> GetNotifications(ReceiverTypes receiver, int userId, int top, int skip)
-        {
-            return await _context.Notifications
-                .Where(n => n.Receiver == receiver && n.UserId == userId)
-                .Take(top)
-                .Skip(skip)
-                .Select(n => new SlimNotification
-                {
-                    Receiver = n.Receiver,
-                    Id = n.Id,
-                    UserId = n.UserId,
-                    Message = n.Message,
-                    Type = n.Type,
-                    Created = n.Created,
-                    IsRead = n.IsRead
-                })
-                .ToListAsync();
-        }
-
-
-        private async Task SendEmail(EmailSettings settings)
-        {
-            if (settings.Emails.Count == 1)
-                await _mailSender.Send(settings.TemplateId, settings.Emails[0], new Api.Models.Mailing.DataWithCompanyInfo());
+            else if (apiCaller.Type == ApiCallerTypes.Admin)
+                return await Send(apiCaller.Id, message, notificationType, emails, templateId);
             else
-                await _mailSender.Send(settings.TemplateId, settings.Emails, new Api.Models.Mailing.DataWithCompanyInfo());
+                return Result.Success();
         }
 
 
-        private async Task SendMessageToAgent(int userId, int? agencyId, int messageId, JsonDocument message)
+        public async Task<Result> Send(int adminId, JsonDocument message, NotificationTypes notificationType, string email = "", string templateId = "")
         {
-            await _agentNotificationHub.Clients
-                .Group($"{agencyId}-{userId}")
-                .ReceiveMessage(messageId, message);
+            return await Send(adminId, message, notificationType, new List<string> { email }, templateId);
         }
 
 
-        private async Task SendMessageToAdmin(int userId, int messageId, JsonDocument message)
+        public async Task<Result> Send(int adminId, JsonDocument message, NotificationTypes notificationType, List<string> emails = null, string templateId = "")
         {
-            await _adminNotificationHub.Clients
-                .Group($"admin-{userId}")
-                .ReceiveMessage(messageId, message);
+            return await _notificationOptionsService.GetNotificationOptions(adminId, ApiCallerTypes.Admin, null, notificationType)
+                .Map(notificationOptions => BuildSettings(notificationOptions, emails, templateId))
+                .Tap(sendingSettings => AddAdminNotification(adminId, message, notificationType, sendingSettings));
         }
 
 
-        private readonly EdoContext _context;
-        private readonly IHubContext<AgentNotificationHub, INotificationClient> _agentNotificationHub;
-        private readonly IHubContext<AdminNotificationHub, INotificationClient> _adminNotificationHub;
-        private readonly MailSenderWithCompanyInfo _mailSender;
-        private readonly IDateTimeProvider _dateTimeProvider;
+        public async Task<Result> Send(SlimAgentContext agent, JsonDocument message, NotificationTypes notificationType, string email = "", string templateId = "")
+        {
+            return await Send(agent, message, notificationType, new List<string> { email }, templateId);
+        }
+
+
+        public async Task<Result> Send(SlimAgentContext agent, JsonDocument message, NotificationTypes notificationType, List<string> emails = null, string templateId = "")
+        {
+            return await _notificationOptionsService.GetNotificationOptions(agent.AgentId, ApiCallerTypes.Agent, agent.AgencyId, notificationType)
+                .Map(notificationOptions => BuildSettings(notificationOptions, emails, templateId))
+                .Tap(sendingSettings => AddAgentNotification(agent, message, notificationType, sendingSettings));
+        }
+
+
+        public async Task<Result> Send(SlimAgentContext agent, DataWithCompanyInfo messageData, NotificationTypes notificationType, List<string> emails = null, string templateId = "")
+        {
+            return await _notificationOptionsService.GetNotificationOptions(agent.AgentId, ApiCallerTypes.Agent, agent.AgencyId, notificationType)
+                .Map(notificationOptions => BuildSettings(notificationOptions, emails, templateId))
+                .Tap(sendingSettings => AddAgentNotification(agent, messageData, notificationType, sendingSettings));
+        }
+
+
+        private static Dictionary<ProtocolTypes, object> BuildSettings(SlimNotificationOptions notificationOptions, List<string> emails, string templateId)
+        {
+            var sendingSettings = new Dictionary<ProtocolTypes, object>();
+
+            if ((notificationOptions.EnabledProtocols & ProtocolTypes.WebSocket) == ProtocolTypes.WebSocket)
+                sendingSettings.Add(ProtocolTypes.WebSocket, new WebSocketSettings { });
+
+            if ((notificationOptions.EnabledProtocols & ProtocolTypes.Email) == ProtocolTypes.Email)
+                sendingSettings.Add(ProtocolTypes.Email, new EmailSettings { Emails = emails ?? new(0), TemplateId = templateId });
+
+            return sendingSettings;
+        }
+
+
+
+
+
+        private readonly IInternalNotificationService _internalNotificationService;
+        private readonly INotificationOptionsService _notificationOptionsService;
+        private readonly IAgentContextService _agentContextService;
     }
 }
